+++ packages/cli/src/commands/components/push/graph-operations/index.ts
@@ -30,7 +30,7 @@ export async function pushWithDependencyGraph(
   space: string,
   spaceState: SpaceComponentsDataState,
 
-  maxConcurrency: number = getActiveConfig().api.rateLimit,
+  backpressure: number = getActiveConfig().api.rateLimit,
 ): Promise<PushResults> {
   // Build and validate the dependency graph with colocated target data
   const context: GraphBuildingContext = { spaceState };
@@ -38,7 +38,7 @@ export async function pushWithDependencyGraph(
   validateGraph(graph);
 
   // Process all resources using the dependency graph
-  const results = await processAllResources(graph, space, maxConcurrency);
+  const results = await processAllResources(graph, space, backpressure);
 
   return results;
 }
